% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{self}
\alias{self}
\title{self}
\usage{
self(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
the param x
}
\description{
this function returns X
}
\details{
DETAILS
}
\examples{
self(2)
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
